backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pandas
groq
datasets
scikit-learn
openai
numpy
python-multipart
